requirements.txt
aiogram
openai
python-dotenv
apscheduler
